Move components into its own directories
